Meta: Changed C++ Compiler g++ to clang to be able to use llvm

+++ src/loop.cpp
@@ -9,7 +9,7 @@ int main(int argc, const char *argv[])
     {
         Error error_handler;
         IO io(error_handler);
-        for (int i = 0; i < argc - 1; i++)
+        for (size_t i = 0; (int)i < argc - 1; i++)
         {
             try
             {
